refactored SearchCriteria, DiscountDAO, DiscountService, added CriteriaDto.

+++ sandbox-backend/src/main/java/com/exadel/sandbox/team5/entity/Review.java
@@ -18,7 +18,7 @@ import java.util.Date;
 public class Review extends BaseEntity {
 
     @Column(name = "rate")
-    private Long rate;
+    private Integer rate;
 
     @Column(name = "comment")
     private String comment;
